pyproject.toml
[project]
name = "heracles_ai"
version = "0.1.0"
description = "Orchestrates personalized fitness and diet programs, suggests helpful resources, and provides support throughout the user's health journey, from initial planning and goal setting to real-time workout guidance and progress tracking."
authors = [
    {name = "Gary Ng", email = "gkcng@google.com"},
    {name = "Arun Santhanagopalan", email = "asanthan@google.com"},
    {name = "Renee Zhang", email = "xiaolizhang@google.com"},
    {name = "Anisha Kaul", email = "anishasjkaul@google.com"},
    {name = "Duncan Renfrow-Symon", email = "drenfrowsymon@google.com"}
]
license = "Apache License 2.0"
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry.dependencies]
google-cloud-aiplatform = { extras = [
    "agent_engines",
    "adk",
], git = "https://github.com/googleapis/python-aiplatform.git", rev = "copybara_738852226" }
python = "^3.11"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-genai = "^1.9.0"
google-adk = ">=0.0.2"
requests = "^2.31.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
google-adk = { version = ">=0.0.2", extras = ["eval"] }

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"
cloudpickle = "^3.1.1"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"